Feat #17430: change swing ratio to use float type for improved customizability (in progress)

+++ src/engraving/rw/write/twrite.cpp
@@ -2811,7 +2811,7 @@ void TWrite::writeProperties(const StaffTextBase* item, XmlWriter& xml, WriteCon
         } else {
             swingUnit = DurationType::V_ZERO;
         }
-        int swingRatio = item->swingParameters().swingRatio;
+        float swingRatio = item->swingParameters().swingRatio;
         xml.tag("swing", { { "unit", TConv::toXml(swingUnit) }, { "ratio", swingRatio } });
     }
